refactor(web/acl): drop legacy page and rename acl-ng to acl
The legacy ACL page consumed a wire-incompatible API surface and was
no longer usable against the current dataplane. The new page shipped
under the acl-ng label is now the only ACL UI.

Delete the legacy page, rename the directory and identifiers from
acl-ng to acl, and collapse the duplicate menu item, route, and API
binding so the sidebar shows a single ACL entry.
