+++ web/src/api/acl.ts
@@ -1,11 +1,11 @@
 import { createService, type CallOptions } from './client';
-import type { IPAddressWire } from '../utils/netip';
 
-// ACL types based on aclpb/acl.proto
+// Types matching aclpb/acl.proto and filterpb/filter.proto exactly.
+// No Action.counter, no keep_state, no MapConfig, no SyncConfig, no DUMP kind.
 
 export interface IPNet {
-    addr?: string | Uint8Array | number[]; // Base64 encoded bytes or raw bytes
-    mask?: string | Uint8Array | number[]; // Base64 encoded bytes or raw bytes
+    addr?: string | Uint8Array | number[];
+    mask?: string | Uint8Array | number[];
 }
 
 export interface PortRange {
@@ -28,111 +28,84 @@ export interface Device {
 }
 
 export enum ActionKind {
-    PASS = 0,
-    DENY = 1,
-    COUNT = 2,
-    CHECK_STATE = 3,
-    CREATE_STATE = 4,
+    ACTION_KIND_PASS = 0,
+    ACTION_KIND_DENY = 1,
+    ACTION_KIND_COUNT = 2,
+    ACTION_KIND_CHECK_STATE = 4,
+    ACTION_KIND_CREATE_STATE = 5,
+    ACTION_KIND_LOG = 6,
 }
+
+/** Human-readable label for each ActionKind. */
+export const ACTION_KIND_LABELS: Record<ActionKind, string> = {
+    [ActionKind.ACTION_KIND_PASS]: 'pass',
+    [ActionKind.ACTION_KIND_DENY]: 'deny',
+    [ActionKind.ACTION_KIND_COUNT]: 'count',
+    [ActionKind.ACTION_KIND_CHECK_STATE]: '?state',
+    [ActionKind.ACTION_KIND_CREATE_STATE]: '+state',
+    [ActionKind.ACTION_KIND_LOG]: 'log',
+};
 
 export interface Action {
     kind?: ActionKind;
-    counter?: string;
-    keep_state?: boolean;
 }
 
 export interface Rule {
-    action?: Action;
-    srcs?: IPNet[];
-    dsts?: IPNet[];
-    src_port_ranges?: PortRange[];
-    dst_port_ranges?: PortRange[];
+    actions?: Action[];
+    counter?: string;
     devices?: Device[];
     vlan_ranges?: VlanRange[];
+    srcs?: IPNet[];
+    dsts?: IPNet[];
     proto_ranges?: ProtoRange[];
+    src_port_ranges?: PortRange[];
+    dst_port_ranges?: PortRange[];
 }
 
-export interface MapConfig {
-    index_size?: number;
-    extra_bucket_count?: number;
-}
-
-export interface SyncConfig {
-    src_addr?: IPAddressWire;
-    dst_ether?: string | Uint8Array | number[];
-    dst_addr_multicast?: IPAddressWire;
-    port_multicast?: number;
-    dst_addr_unicast?: IPAddressWire;
-    port_unicast?: number;
-    tcp_syn_ack?: number | string; // nanoseconds
-    tcp_syn?: number | string;
-    tcp_fin?: number | string;
-    tcp?: number | string;
-    udp?: number | string;
-    default?: number | string;
-}
-
-// Request/Response types
-
-export interface AclUpdateConfigRequest {
-    name?: string;
-    rules?: Rule[];
-}
-
-export interface AclUpdateConfigResponse { }
-
-export interface AclShowConfigRequest {
-    name?: string;
-}
-
-export interface AclShowConfigResponse {
-    name?: string;
-    rules?: Rule[];
-    fwstate_map?: MapConfig;
-    fwstate_sync?: SyncConfig;
-}
-
-export interface AclDeleteConfigRequest {
-    name?: string;
-}
-
-export interface AclDeleteConfigResponse {
-    deleted?: boolean;
-}
-
-export interface AclListConfigsRequest { }
-
-export interface AclListConfigsResponse {
+export interface ListConfigsResponse {
     configs?: string[];
 }
 
-export interface AclUpdateFWStateConfigRequest {
+export interface ShowConfigRequest {
     name?: string;
-    map_config?: MapConfig;
-    sync_config?: SyncConfig;
 }
 
-export interface AclUpdateFWStateConfigResponse { }
+export interface ShowConfigResponse {
+    name?: string;
+    rules?: Rule[];
+    fwstate_name?: string;
+}
+
+export interface UpdateConfigRequest {
+    name?: string;
+    rules?: Rule[];
+}
+
+export interface UpdateConfigResponse {}
+
+export interface DeleteConfigRequest {
+    name?: string;
+}
+
+export interface DeleteConfigResponse {
+    deleted?: boolean;
+}
 
 const aclService = createService('aclpb.ACLService');
 
 export const acl = {
-    listConfigs: (options?: CallOptions): Promise<AclListConfigsResponse> => {
-        return aclService.call<AclListConfigsResponse>('ListConfigs', options);
-    },
-    showConfig: (request: AclShowConfigRequest, options?: CallOptions): Promise<AclShowConfigResponse> => {
-        return aclService.callWithBody<AclShowConfigResponse>('ShowConfig', request, options);
-    },
-    updateConfig: (request: AclUpdateConfigRequest, options?: CallOptions): Promise<AclUpdateConfigResponse> => {
-        return aclService.callWithBody<AclUpdateConfigResponse>('UpdateConfig', request, {
+    listConfigs: (options?: CallOptions): Promise<ListConfigsResponse> =>
+        aclService.call<ListConfigsResponse>('ListConfigs', options),
+
+    showConfig: (request: ShowConfigRequest, options?: CallOptions): Promise<ShowConfigResponse> =>
+        aclService.callWithBody<ShowConfigResponse>('ShowConfig', request, options),
+
+    updateConfig: (request: UpdateConfigRequest, options?: CallOptions): Promise<UpdateConfigResponse> =>
+        aclService.callWithBody<UpdateConfigResponse>('UpdateConfig', request, {
             compress: true,
             ...options,
-        });
-    },
-    deleteConfig: (request: AclDeleteConfigRequest, options?: CallOptions): Promise<AclDeleteConfigResponse> => {
-        return aclService.callWithBody<AclDeleteConfigResponse>('DeleteConfig', request, options);
-    },
-    updateFWStateConfig: (request: AclUpdateFWStateConfigRequest, options?: CallOptions): Promise<AclUpdateFWStateConfigResponse> => {
-        return aclService.callWithBody<AclUpdateFWStateConfigResponse>('UpdateFWStateConfig', request, options);
-    },
+        }),
+
+    deleteConfig: (request: DeleteConfigRequest, options?: CallOptions): Promise<DeleteConfigResponse> =>
+        aclService.callWithBody<DeleteConfigResponse>('DeleteConfig', request, options),
 };
