+++ web/src/api/index.ts
@@ -7,8 +7,7 @@ export * from './functions';
 export * from './pipelines';
 export * from './devices';
 export * from './decap';
-export * from './acl';
-export { aclng, ACTION_KIND_LABELS } from './acl-ng';
+export { acl, ACTION_KIND_LABELS } from './acl';
 export * from './forward';
 export * from './counters';
 
@@ -20,7 +19,6 @@ import { pipelines } from './pipelines';
 import { devices } from './devices';
 import { decap } from './decap';
 import { acl } from './acl';
-import { aclng } from './acl-ng';
 import { forward } from './forward';
 import { counters } from './counters';
 
@@ -34,7 +32,6 @@ export const API = {
     devices,
     decap,
     acl,
-    aclng,
     forward,
     counters,
 };
