--- web/src/pages/modules/acl-ng/index.ts
@@ -1,1 +1,0 @@
-export { default as AclNgPage } from './AclNgPage';
