--- web/src/pages/modules/acl/VirtualizedAclTable.scss
@@ -1,119 +1,0 @@
-/* ============================================
-   VirtualizedAclTable Styles
-   ============================================ */
-
-.acl-table {
-    display: flex;
-    flex-direction: column;
-}
-
-.acl-table__container {
-    height: 100%;
-}
-
-/* Search Bar */
-.acl-table__search-bar {
-    display: flex;
-    align-items: center;
-    gap: 16px;
-    flex-shrink: 0;
-}
-
-.acl-table__search-input {
-    width: 350px;
-}
-
-.acl-table__stats {
-    display: flex;
-    align-items: center;
-    gap: 8px;
-}
-
-/* Table Container */
-.acl-table__wrapper {
-    flex: 1;
-    border: 1px solid var(--g-color-line-generic);
-    border-radius: 8px;
-    overflow: hidden;
-    display: flex;
-    flex-direction: column;
-    position: relative;
-}
-
-/* Loading Overlay */
-.acl-table__loading-overlay {
-    position: absolute;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    background-color: var(--g-color-base-background);
-    opacity: 0.7;
-    z-index: 10;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-/* Table Header */
-.acl-table__header {
-    display: flex;
-    align-items: center;
-    padding: 0 8px;
-    border-bottom: 1px solid var(--g-color-line-generic);
-    background-color: var(--g-color-base-generic);
-    font-weight: 500;
-    box-sizing: border-box;
-}
-
-/* Table Body */
-.acl-table__body {
-    overflow: auto;
-    contain: strict;
-}
-
-.acl-table__virtual-container {
-    width: 100%;
-    position: relative;
-}
-
-/* Table Row */
-.acl-table__row {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    display: flex;
-    align-items: center;
-    padding: 0 8px;
-    border-bottom: 1px solid var(--g-color-line-generic);
-    box-sizing: border-box;
-}
-
-.acl-table__row--even {
-    background-color: transparent;
-}
-
-.acl-table__row--odd {
-    background-color: var(--g-color-base-generic-ultralight);
-}
-
-/* Empty State */
-.acl-table__empty {
-    padding: 40px 20px;
-    text-align: center;
-}
-
-/* Footer */
-.acl-table__footer {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    flex-shrink: 0;
-}
-
-/* Action Text */
-.acl-table__action-text {
-    font-weight: 500;
-}
-
