--- web/src/pages/modules/acl/dialogs/index.ts
@@ -1,3 +1,0 @@
-export { UploadYamlDialog } from './UploadYamlDialog';
-export { CreateConfigDialog } from './CreateConfigDialog';
-export { UnsavedChangesDialog } from './UnsavedChangesDialog';
