+++ web/src/types.ts
@@ -8,7 +8,6 @@ export const PAGE_IDS = [
     'modules/forward',
     'modules/decap',
     'modules/acl',
-    'modules/acl-ng',
     'modules/pdump',
     'modules/route',
     'operators/route',
